Refactor UI wiring, mounting, and scheduler modules
