+++ cannaclicker/src/app/ui/events/scheduler/index.ts
@@ -1,0 +1,9 @@
+import { createScheduler } from "./scheduler";
+import { defaultSchedulerBindings } from "./uiBindings";
+import type { SchedulerContext } from "./types";
+
+export type { SchedulerContext, SchedulerBindings } from "./types";
+
+export function createEventScheduler(context: SchedulerContext) {
+  return createScheduler(context, defaultSchedulerBindings);
+}
