+++ cannaclicker/src/app/ui/types/abilities.ts
@@ -1,0 +1,12 @@
+export interface AbilityButtonRefs {
+  container: HTMLButtonElement;
+  icon: HTMLImageElement;
+  label: HTMLElement;
+  status: HTMLElement;
+  progressBar: HTMLElement;
+}
+
+export interface UIAbilityPanelRefs {
+  abilityTitle: HTMLElement;
+  abilityList: Map<string, AbilityButtonRefs>;
+}
