+++ cannaclicker/src/app/ui/types/clicker.ts
@@ -1,0 +1,6 @@
+export interface UIClickerRefs {
+  clickButton: HTMLButtonElement;
+  clickLabel: HTMLSpanElement;
+  clickIcon: HTMLDivElement;
+  announcer: HTMLElement;
+}
