+++ cannaclicker/src/app/ui/types/controls.ts
@@ -1,0 +1,14 @@
+export interface ControlButtonRefs {
+  button: HTMLButtonElement;
+  icon: HTMLImageElement;
+  label: HTMLSpanElement;
+}
+
+export interface UIControlGroupRefs {
+  controls: {
+    mute: ControlButtonRefs;
+    export: ControlButtonRefs;
+    import: ControlButtonRefs;
+    reset: ControlButtonRefs;
+  };
+}
