+++ cannaclicker/src/app/ui/types/header.ts
@@ -1,0 +1,3 @@
+export interface UIHeaderRefs {
+  headerTitle: HTMLHeadingElement;
+}
