+++ cannaclicker/src/app/ui/types/seeds.ts
@@ -1,0 +1,4 @@
+export interface UISeedRefs {
+  seedBadge: HTMLButtonElement;
+  seedBadgeValue: HTMLElement;
+}
