+++ cannaclicker/src/app/ui/types/services.ts
@@ -1,0 +1,6 @@
+export interface UIServiceRefs {
+  toastContainer?: HTMLElement;
+  eventLayer?: HTMLElement;
+  modalOverlay?: HTMLElement;
+  eventRoot?: HTMLElement;
+}
