+++ cannaclicker/src/app/ui/types/stats.ts
@@ -1,0 +1,11 @@
+export interface UIStatRefs {
+  statsLabels: Map<string, HTMLElement>;
+  statsMeta: Map<string, HTMLElement>;
+  buds: HTMLElement;
+  bps: HTMLElement;
+  bpc: HTMLElement;
+  total: HTMLElement;
+  seeds: HTMLElement;
+  seedRate: HTMLElement;
+  prestigeMult: HTMLElement;
+}
